Reduce selection, refactor
