+++ src/libs/Clipboard/index.native.js
@@ -3,7 +3,7 @@ import Clipboard from '@react-native-community/clipboard';
 export default {
     ...Clipboard,
 
-    // Native does not support HTML.
+    // We don't want to set HTML on native platforms so noop them.
     canSetHtml: () => false,
     setHtml: () => {},
 };
