+++ src/libs/SelectionScraper/index.native.js
@@ -1,8 +1,4 @@
-import BaseSelectionScraper from './BaseSelectionScraper';
-
 export default {
-    ...BaseSelectionScraper,
-
     // This is a no-op function for native devices because they wouldn't be able to support Selection API like a website.
     getCurrentSelection: () => {},
 };
